+++ covasim/base.py
@@ -566,6 +566,34 @@ class BasePeople(sc.prettyobj):
         self.init_contacts() # Initialize the contacts
         self.infection_log = [] # Record of infections - keys for ['source','target','date','layer']
 
+        # Set person properties -- all floats except for UID
+        for key in self.meta.person:
+            if key == 'uid':
+                self[key] = np.arange(self.pop_size, dtype=cvd.default_int)
+            else:
+                self[key] = np.full(self.pop_size, np.nan, dtype=cvd.default_float)
+
+        # Set health states -- only susceptible is true by default -- booleans
+        for key in self.meta.states:
+            if key == 'susceptible':
+                self[key] = np.full(self.pop_size, True, dtype=bool)
+            else:
+                self[key] = np.full(self.pop_size, False, dtype=bool)
+
+        # Set dates and durations -- both floats
+        for key in self.meta.dates + self.meta.durs:
+            self[key] = np.full(self.pop_size, np.nan, dtype=cvd.default_float)
+
+        # Store the dtypes used in a flat dict
+        self._dtypes = {key:self[key].dtype for key in self.keys()} # Assign all to float by default
+        self._lock = True # Stop further keys from being set (does not affect attributes)
+
+        # Store flows to be computed during simulation
+        self.flows = {key:0 for key in cvd.new_result_flows}
+
+        # Although we have called init(), we still need to call initialize()
+        self.initialized = False
+
         return
 
 
@@ -665,12 +693,29 @@ class BasePeople(sc.prettyobj):
         return (self[key]==0).sum()
 
 
-    def keys(self, which=None):
-        ''' Returns the name of the states '''
-        if which is None:
-            return self.meta.all_states[:]
-        else:
-            return getattr(self.meta, which)[:]
+    def keys(self):
+        ''' Returns keys for all properties of the people object '''
+        return self.meta.all_states[:]
+
+
+    def person_keys(self):
+        ''' Returns keys specific to a person (e.g., their age) '''
+        return self.meta.person[:]
+
+
+    def state_keys(self):
+        ''' Returns keys for different states of a person (e.g., symptomatic) '''
+        return self.meta.states[:]
+
+
+    def date_keys(self):
+        ''' Returns keys for different event dates (e.g., date a person became symptomatic) '''
+        return self.meta.dates[:]
+
+
+    def dur_keys(self):
+        ''' Returns keys for different durations (e.g., the duration from exposed to infectious) '''
+        return self.meta.durs[:]
 
 
     def layer_keys(self):